+++ docs/conf.py
@@ -171,5 +171,7 @@ texinfo_documents = [
      'Miscellaneous'),
 ]
 
+def setup(app):
+    app.add_stylesheet('css/custom.css')  # may also be an URL