+++ src/ontask/urls.py
@@ -89,4 +89,11 @@ handler404 = 'ontask.views.ontask_handler404'
 handler500 = 'ontask.views.ontask_handler500'
 
 # Create the DB engine with SQLAlchemy (once!)
-pandas_db.engine = pandas_db.create_db_engine()
+pandas_db.engine = pandas_db.create_db_engine(
+    'postgresql',
+    '+psycopg2',
+    settings.DATABASES['default']['USER'],
+    settings.DATABASES['default']['PASSWORD'],
+    settings.DATABASES['default']['HOST'],
+    settings.DATABASES['default']['NAME'],
+)